Fix agent crash at FlowStatsManager::FreeReqHandler
Issue:
FlowStatsCollectorObject can have multiple FlowStatsCollector instances. When
the first instance of FlowStatsCollector is ready for delete we are attempting
to delete the FlowStatsCollectorObject itself without verifying whether other
instances of FlowStatsCollector are ready to be deleted.

Fix:
Delete FlowStatsCollectorObject only when all instances of its
FlowStatsCollector are ready to be deleted.

Change-Id: Ib2abafcdc01b1587634b9a2586228cd0f7204041
Closes-Bug: #1650505