Ensure flow-stickiness in case of ECMP with bridged forward flow and routed reverse flow
Commit 14333196c7fe7f2351a9552521719f979373c102 supported stickiness
across bridge and routed flows. However, it doenst support multiple ECMP
members on same compute node.

This current supports multiple members of ECMP in same compute

When layer-3 flow is created with local-ecmp-nh as key, look for layer-2 flow
with one of the local-ecmp members as key. If layer-2 flow with interface-nh
is present, stitch layer-3 and layer-2 flows

Conflicts:
	src/vnsw/agent/pkt/flow_entry.cc
	src/vnsw/agent/pkt/flow_entry.h
	src/vnsw/agent/pkt/test/egress-flow.xml
	src/vnsw/agent/pkt/test/test_pkt_util.cc
	src/vnsw/agent/pkt/test/test_pkt_util.h

Change-Id: I3fd0185fb9856fbcbdc2b038a75fa65d54183802
closes-Bug: #1648696