Ensure flow-stickiness in case of ECMP with bridged forward flow and routed reverse flow
Commit 14333196c7fe7f2351a9552521719f979373c102 supported stickiness
across bridge and routed flows. However, it doenst support multiple ECMP
members on same compute node.

This current supports multiple members of ECMP in same compute

When layer-3 flow is created with local-ecmp-nh as key, look for layer-2 flow
with one of the local-ecmp members as key. If layer-2 flow with interface-nh
is present, stitch layer-3 and layer-2 flows

Change-Id: I0ba4718d212639413cba257657605a09bcabd778
Closes-Bug: #1648696