Resource manager
Role of this infrastructure is to manage various kind of resources used by agent
and make it persistent by providing a backup method.

Currently the resource being pushed in is index_vector. However any other kind
can also be used in same.
For each resource there can be multiple incarnation. Like index vector can have
one incarnation for MPLS labels while other for nexthop index.
These incarnations can then have multiple users.
Taking the example of MPLS which is being pushed in current check-in, is index
vector type and can be used by modules like VMI, route, VRF, Multicast or even
test.

There is a backup module which maps all the resources to a sandesh map and flush
it on to file to make it persistent. On restart it reads the same set of files
back and build the resources back again. Now when config is processed and oper
entries come into existence, they ask for resource which will be similar to the
one allocated before.

Each client/user is identified by a key which is backed up as well. Say VRF
asked for label before restart, so its key will be vrf-name and some index 100
will be given. Same information mapping VRF name to index 100 is backed up. On
restart again this VRF gets created and it asks for resource using key as
vrf-name. Since resource manager has read back the same from file it gives 100
back to VRF.

Note: Backing up will be made optional. Resource manager is not but whether to
persist data will be made optional.

Partial-bug: #1654481
Change-Id: I8795e48cbaf00a81e08bf96d37876cfd029c5081