Node manager fixes to work with systemd
Changes to make node manager work with both systemd
and supervisord. Abstracted the process information
manager portion of node manager so that it can work
with both systems. Refactor EventManager so that the
node type specific code is abstracted out. Change
supervisord event TICK_60 to a periodic instead so
that common code can be used between supervisord and
systemd.

The systemd process info manager code uses pydbus
which will be installed as part of another commit.

Change-Id: I49a2b5f505bc7edc8e87c8facb8343f7f7c3dc56
Partial-Bug: #1651777