* Handle delete of mac mgmt entry if db_entry is not present
If a PBB tunnel packet points to a BMAC which is not
present in route table and the mac entry gets aged out
then while deleting the mac entry dont access the
db_entry as it will not be present.
Also removing deletion of this pointer in mac mgmt code.

Change-Id: Ia1832b88e8c07bc7f0a76afe59a5c65db7967c66
Closes-bug:#1655693