Fixed connection issue with kube_api_server
kube_api_server closes the connection which is registered
for watch event on timeout which caused the issue.
By increasing the timeout in the kube_api_server,
the connection close can be postponed. But it
wont solve the issue completly. fix added to reinitiate
the connection, if it is lost.

Change-Id: Ic3318588d6176e6e7a977800c625ed18187e2d2e
Closes-bug: 1657918