Allow prefix string in route introspect to be a perl-style regex
The specified prefix string is treated as a regex and the string
representation of each route prefix searched for the regex.  Note
that the regex need not match the entire string representation of
a prefix - it only has to match a sub-sequence.

Shorter and longer matches are not supported when using a regex.

Change-Id: Ie6c8db700524b4fd02f3f4ae36c5dd91c0ce4fc0
Closes-Bug: 1663752
(cherry picked from commit 5cca03efd678b300c85721940e18a30406a06147)