Vm project UUID should be optional for Port Add requests.
Currently the REST interface of contrail-vrouter-agent for Port Adds, makes it mandatory for vm-project-uuid to be specified when the port type is NovaVMPort. In case of thrift interface this was not mandatory. To ensure backward compatibility, vm-project-uuid should be optional even in REST interface.
Also change the default value for port-type in vrouter_api.py to NameSpacePort to ensure backward compatibility.

Change-Id: I04620fe9b2997c39e46d3069cef8fb82d9f22270
Closes-Bug: #1664819