Vm project UUID should be optional for Port Add requests.
Currently the REST interface of contrail-vrouter-agent for Port Adds, makes it mandatory for vm-project-uuid to be specified when the port type is NovaVMPort. In case of thrift interface this was not mandatory. To ensure backward compatibility, vm-project-uuid should be optional even in REST interface.
Also change the default value for port-type in vrouter_api.py to NameSpacePort to ensure backward compatibility.

Closes-Bug: #1664819
(cherry picked from commit de72bafd363f96048ece41785b0b6f55e64f3a0d)

Change-Id: Iacad9fc68acd6affcf9f8c3908644924407e8077