Vm project UUID should be optional for Port Add requests.
Currently the REST interface of contrail-vrouter-agent for Port Adds, makes it mandatory for vm-project-uuid to be specified when the port type is NovaVMPort. In case of thrift interface this was not mandatory. To ensure backward compatibility, vm-project-uuid should be optional even in REST interface.
Also change the default value for port-type in vrouter_api.py to NameSpacePort to ensure backward compatibility.

Change-Id: Ifc434992f87d8d8cff2da169d6bfeb308b9903bb
Closes-Bug: #1664819