* Stick new flows and mark only the dangling flow as short flow
When flow destination transitions from local compute to remote
compute or viceversa, key of the new flow would be different from
the existing flow. In current code all the 3 flows were marked as
short flow, change it such that latest correct flows are not
short flow and only invalid flow is marked short flow and deleted.

Change-Id: If0cd6879ed0ef61d98bbe240efef8360b0922f7e
Closes-bug:#1665557