Fix race condition in BgpPeer creation for non-master instances
Two problems related to BgpPeer creation in non-master instances
are fixed.

1. If a new incarnation of a routing-instance and neighbors in the
routing-instance are configured before the previous incarnation of
the routing-instance has been destroyed, we end up trying to create
a neighbor that tries to take a LifetimeRef to a deleted instance.

2. When the previous incarnation of a routing-instance is destroyed
and a new incarnaiton is created, also create neighbors that belong
to the routing-instance.

Change-Id: Ic85893ed301b9221e07dfd740974b8e925835bb6
Closes-Bug: 1668369