Fix race condition in BgpPeer creation for non-master instances
If a new incarnation of a routing-instance and neighbors in the
routing-instance are configured before the previous incarnation of
the routing-instance has been destroyed, we end up trying to create
a neighbor that tries to take a LifetimeRef to a deleted instance.

Ignore neighbor create/update notifications if the routing-instance
is marked deleted. The neighbors under the routing-instance will get
created when the instance itself gets resurrected.

Change-Id: I93d74cf8f903fa8f6a1949c990a70481d654ecb3
Closes-Bug: 1668369