Do not return id_perms for list request by admin user
Currently, as part of list operation, we read the id_perms field
to make sure the user has permission to read the object. However,
when returning the list back, the field is omitted. But for admin
user, this step is skipped and id_perms field is being returned.

With this code change, we avoid even reading the id_perms field
for admin users.

Change-Id: I3c775c30873c5ca69abe8c339d598d6f6c21aec7
Closes-Bug: 1668385