Remove an invalid assert over cql query result
db multi-get rows with "in" clause no longer ensures that missing rows get
empty output. Hence remove associated assert on that condition

In cases, especially when we migrate from old to new ones, it is possible that
some of the uuids are missing from the obt_to_dict table

In such cases, log a warning message and delete associated row entirely.
Also log a messge when type/fq_name is not found for any row object.

Change-Id: Ia15af4810cfd45d3f78a0200f0ea862fa6f01141
Partial-Bug: #1632470