Replace obj_dict field in rabbitmq create message with just fq_name
Currently, we populate the entire obj_dict in rabbitmq create and delete
messages. For create, there is no need to send the entire object. Just
fq name is sufficient because it is used by control node and schema
transformer to build their cache of uuid to fq_name mapping.

We will retain the obj_dict for delete messages because there is no way
for the clients to read the data if needed.

Also streamlined some of the internal methods to take just a uuid
instead of a dict containing a uuid.

Change-Id: If47ed704b50669569ef24d302fc805f0e233f480
Closes-Bug: 1670530