* Stick new flows and mark only the dangling flow as short flow
When flow destination transitions from local compute to remote
compute or viceversa, key of the new flow would be different from
the existing flow. In current code all the 3 flows were marked as
short flow, change it such that latest correct flows are not
short flow and only invalid flow is marked short flow and deleted.

Change-Id: If24c2fc2a5fb9e9098e35b2a075c8de445d17182
Closes-bug:#1665557