Patch for IP pool issues:
The problem is when we have more than one api-server, when a subnet is
deleted, the other api-servers keep the subnet allocator in their
subnet_allocator dict (in vnc_db.py). So when we create a subnet with
the same parameters (net, name, cidr,...) when we attach a port to him,
the allocated api is not always in the good range, depending on which
api-server responded to the create-port request.

- net_(delete/update)_notify methods modified,to update dicts on each nodes.

- ipam_(delete/update)_notify modified too.

- new parameter added (notify) now, "IndexAllocator.delete_all"
  is done only from "*_req" methods.

Change-Id: I98f095aeca37290c98bf12fb2be57e74b66efa0a
Closes-Bug: #1332645
Closes-Bug: #1365824