Patch for IP pool issues:
The problem is when we have more than one api-server, when a subnet is
deleted, the other api-servers keep the subnet allocator in their
subnet_allocator dict (in vnc_db.py). So when we create a subnet with
the same parameters (net, name, cidr,...) when we attach a port to him,
the allocated api is not always in the good range, depending on which
api-server responded to the create-port request.

- new parameter "notified" added to delete_subnet_allocator method to delete
  allocator from subnet_allocator dict and not the IndexAllocator from
  ZK (it has already been deleted from the api-serve that received the
  delete request).

- net_(delete/update)_(req/notify) methods modified.
  now, "req" methods delete data from ZooKeeper while
  _subnet_allocators's elements (vnc_db.py) are deleted from "notify"
  methods. _subnet_objs (vnc_addr_mgmt.py) are also deleted from
  "notify" methods

- ipam_(delete/update)_(req/notify) updated to match the new model

- "vnc_rdbms.py" file updated

Change-Id: I98f095aeca37290c98bf12fb2be57e74b66efa0a
Closes-Bug: #1332645
Closes-Bug: #1365824