Patch for IP pool issues:
The problem is when we have more than one api-server, when a subnet is
deleted, the other api-servers keep the subnet allocator in their
subnet_allocator dict (in vnc_db.py). So when we create a subnet with
the same parameters (net, name, cidr,...) when we attach a port to him,
the allocated api is not always in the good range, depending on which
api-server responded to the create-port request.

- "net_update_notify" method updated to delete subnet allocators from
  subnet_allocators dict (in vnc_db.py module) from all api-servers

- subnet_allocators dict modified (in vnc_db.py) to sort subnet by virtual networks
   -->  original dict contained subnet allocators, now the dict contains dicts
        with vn_fq_name as key. So when we need to delete allocator from
        subnet_allocators dict we just parse current vn's allocators.

- added method to get subnet_allocators using vn_fq_name (in vnc_db.py)

- new parameter added to delete_subnet_allocator method to delete
  allocator from subnet_allocator dict and not the IndexAllocator from
  ZK (it has already been deleted from the api-serve that received the
  delete request)

Change-Id: I98f095aeca37290c98bf12fb2be57e74b66efa0a
Closes-Bug: #1332645
Closes-Bug: #1365824