[VNC API Server] Fix UVE sending method on AMQP notifications
The method 'dbe_uve_trace' signature of class vnc_db.VncDbClient
was updated [1] but the same method of class vnc_db.VncServerKombuClient
which depends on it was not.
That fix simplifies it to directly calls
vnc_db.VncDbClient.dbe_uve_trace with AMQP message dict which have the
same signature as the method.
Also remove unused methods and factorize notification methods in
vnc_db.VncServerKombuClient class.

[1] https://review.opencontrail.org/#/c/29589/

Change-Id: I14f74952b5bab3494be23930aa840ff8aa3e7247
Closes-Bug: #1673802