[VNC config] Purge remaining AMQP queue
Config services create AMQP queue to consume messages sent by API
server. When they abort, that queue is not deleted and continue to be
filled. Habitually, these services are directly restarted and put in
standby mode (another node should have take the lead). So the queue will
be purge when the service passes active.
That patch proposes to delete the queue if it exists before the service
goes in standby.

Change-Id: Ie1eb1f2a55d879c712d2914964ee1ff0edba9864
Closes-Bug: #1669698