NH allocation for Ecmp path
Earlier NH allocation for ecmp path happens inside ModifyEcmpPath
api, whereas with changes done for #1664483, composite-NH allocation
happens prior to calling ModifyEcmpPath api. This is done for existence
of MplsLabel(which points to composite-nh) to pass for ModifyEcmpPath api.

Reverting to old way and allocating MplsLabel as well inside the api if
required.

Change-Id: Ie4806279c6796559f2b5404133122df8def606bf
Closes-Bug: #1664483