Fix test_flow_mgr_instances failure
Issue:
------
upon interface delete, route delete and notification is done
before triggering interface notification. which due to multi
level task enqueue and processing results in different order
of events based on chance of processing for a task, further
more flow tokens adds another level of variance to order of
events.

So if route delete is processed before processing interface
delete for a flow, it converts to short flow due to
interface unavailability and since stats collector is not
running to age short flows, it remains stuck in flow_tree
causing test case to fail

Fix:
----
- wait for flow count to reach 200 while flow add
- wait for flow mgmt to complete building db table
  dependency
- before triggering delete of interface, disable flow update
  queue to ensure processing of delete interface to complete
  before flow table being updated for recompute due to route
  delete notification
- use config file with large token values avoid task defer
  due to unavailability of tokens

Change-Id: I4d29fa08f438c40d938b32108fd5163b11e7d976
Closes-Bug: 1675711