Fixes: Routeleak not happening in case of IPv6-SVC with ECMP
The IPv6 Address gets assigned after the SI is created sometimes
causing the si_info not to be populated.

The fix is to check for the SI info even if the SC is created
and recreate SC object if it is different.

UT Added for checking the IPv6 address in the service chain.

Change-Id: I9a020d553853ead0e8e2e90c856ab3f8575961df
Closes-Bug: 1670320