Use TOR IP to build RD for Inclusive Multicast route
Use TOR IP instead of the TOR agent IP to build the RD for the Inclusive
Multicast route for a TOR agent.  Since both the TOR agents use the same
VRF id after the fix for bug 1652162, the IM routes advertised on behalf
of both TOR agents will now have the same RD. Further, since the tag and
IP address of the TOR are the same for both routes already, this results
in the same IM route prefix.

This in turn allows both primary and backup TOR agents to advertise IM
routes without causing traffic duplication of BUM traffic in the steady
state.  This reduces BUM traffic loss in the event of OVSDB connection
flap.

Note that the agent code currently advertises unicast routes from both
primary and backup, but doesn't do so for multicast due to the risk of
traffic duplication. Agent code will be modified appropriately once this
change is merged.

Change-Id: I82ef1d44c4a4b78f64068748c760957aeb66a1d8
Partial-Bug: 1683723