Fix health-check when fat flow enabled on vmi
Problem:
Fat flow is enabled on the vm-interface and not on vhost0 interface.
So, packets from vhost0 result in non-fat flow and packets from
vm-interface result in fat-flow.

Health Check assumes that flows are always initiated from vhost0
interface. So, flows created due to packets from vm-interface do not
have NAT translation based on health-check rules. As a result, packets
from vm-interface are going to host-os

Solution:

1. Health-Check status is ACTIVE
When fat-flow is enabled, the fat-flow created due to packet from
vm-interface will also apply health-check rules and create appropriated
NAT-ed reverse flow.

2. Health-Check status is INACTIVE
When health-check status is INACTIVE, the routes for IP address on
vm-interface are also withdrawn. So, flows cannot be setup for packets
from vm-interface.
As a solution, we disable fat-flow on the vm-interface when health-check
status is INACTIVE. As a result, packets from vm-interface will match
flows created due to request and health-check session will succeed.

Change-Id: I7766cbd9bbc25e1f740148c35ed7e115a4fd7d95
Fixes-Bug: #1676721