Remove the UUID proces list
In the current code, UUID request-list or process-list is stored in two container.
1. List - Ensure the order of read/update
2. SET - To supress duplicate insertion of UUID to UUID process list.

Currently, since the UUID reading is done in multiple partition, the ordering of
read is not ensured. Code does handle the out of order reading of UUID in the
config reader task.

Remove the need for list(time ordered list).

Change-Id: Ia4734b3da90d58e7ab3f4c1bbe870f6f86b297da
Related-bug: #1632470