Add specific unit to test ShowRouteHandler::ConvertReqIterateToReq()
code-coverage indicated clearly that all of the code in this routine is untested
Added a specific test to cover the same. Line coverage for bgp_show_route.c
increased from 72.5% to ~90% with this additional unit test.

Move ShowRouteHandler class definition to a new file and fix a few nits there by
found by cpplint

Closes-Bug: 1681953
Change-Id: I269cd426fe19e4f99a33e2fa1a60f0bca22635d9