Fix port-update API to support mac-address update
If mac address of a port is chnaged via port-update API,
it is ignored. This fix handles the updating of the mac
address. Also, it ensures that the same mac address
does not exixt on any other port in the same virtual
network.

Change-Id: Iec9290e0bc4485f2d59c38b223ddc458679aa1f4
Closes-Bug: 1682672
Partial-Bug: 1683547
(cherry picked from commit 3905546ac6172f001d1d214dbcdc6541c788ed29)