When sandesh generator is initialized as part
of the ConfigServiceLogger initialization, introspect
port of config services are opened and thus all the
config services in all the nodes are marked active,
even is the service did not get the zookeeper lock.

Earlier the ConfigServiceLogger was initialized after
zookeeper lock, Now as part the fix for the following
LP: https://bugs.launchpad.net/juniperopenstack/+bug/1669698,
We are creating ConfigServiceLogger object even before the
zookeeper lock.

Fix is to initialize sandesh generator without spawning the
introspect thread and once the lock is acquired sandesh
generator is reinitialized to spawn the introspect thread.

Change-Id: I8507ed1e7b49fb6cae3570db5612978825940cce
Closes-Bug: 1682683