Fixed floaing_ip update event issue in reaction_map.py in svc_monitor
since the reaction_map for floating_ip and vmi is not there,
floating_ip events are not evaluted. Fixed the reaction_map to handle
the above issue.
floating_ip_fixed_ip_address will be set, if the fip is associated to
a vmi for the public access. It has to be set to None for ecmp
loadbalancer

Change-Id: Ifadf44e701d19c1b858800d6736947b19704f599
Partial-bug: #1639062