HA stale for mcast in tor-agent(TA)
When TA connection with tor goes down because of flap on remote end, multicast
route was withdrawn. This used to impact mcast traffic.
Expectation is to hold on till other TA either picks up tor or connection is
established again with same TA. This is in line with unicast route.
Since TA HA is a listener on bridge table and local_mcast publishes route in
evpn table, this change enables adding of l2 mcast path derived from evpn mcast
route. Previous to this change it was disabled.

Change-Id: If4f74590e6d5bda8920214e1a9acb7cd6739c1ee