Support audit of bridge entries in contrail-vrouter-agent
Vrouter bridge entries are visited and entries which are present only in vrouter are added to audit_list of bridge entries.
The entries which are in this audit_list for more than a given time and still not added by agent are the entries which are selected for removal from vrouter.

Defined a new KSyncEntry for bridge routes which sends message to vrouter only for deletes. The existing KSyncEntry for bridge routes cannot be used because it is tied to DBEntry.
Added UT to verify audit of bridge entries.

Change-Id: I2d03a33f0f452373a75798fe2a02eeba43f6495e
Closes-Bug: #1685249