Making use of the existing commandline args of
the contrail-compute-setup script to populate
the HYPERVISOR section in the contrail-vrouter-agent.conf

contrail-compute-setup is intended to be used with
/etc/contrailctl/agent.conf as input file in agent contianer
and with commandline args in a baremetal compute.

So reading the mode and type from the HYPERVISOR section
of the /etc/contrailctl/agent.conf and populating it as
default value in the contrail-compute-setup and used.

Change-Id: I505a5d4feebe561f6ce5d3c6e0c072da5aafda0f
Partial-Bug: 1683883