Optimize RibOutAttr constructor for BGP RibOuts
There's no need to allocate a NextHop and add it to the NextHopList for
RibOutAttrs in BGP RibOuts. We can simply store the label and l3_label
in the RibOutAttr itself.  All the other attributes are already in the
BgpAttrPtr as there's no need to store ECMP NextHops for BGP RibOuts.

Change-Id: I07ad07165e34d4942f219b7acb28487682830065
Partial-Bug: 1687537