Send a rabbitmq message for both sides of link when a symmetric link is updated
For any link update, API server only send the notification for the object
that was requested to be updated by the user. However, when a symmetric
link (link between objects of same types, E.g. routing-instance to
routing-instance link) is updated, the clients are expected to treat it
similar to both sides being updated.

However, because of control-node implementation, it does not process it
correctly. Fixing this in control-node is difficult, so we are
introducing a new message oper-type for rabbitmq messages. These messages
will be ignored by all other rabbitmq clients except for control-node,
which will use it to update the other side of the link. A subsequent change
in control-node is required for the fix to be complete.

Partial-Bug: 1686901

Change-Id: I5c89c112b86106ea7d3c519cc316f59780d80e85