Used resource lost on audit.
This was seen with mpls labels.
When backup is enabled and agent restarted, the resource restored are stored
with dirty bit. Audit is started after some time. If any user consumes it, this
bit is reset. Audit removes all entry with bit set.
Since reset of bit was not working, even on consumption, resource was removed by
audit. This used to result in resource allocated to some other user, in turn
having multiple users contending for same resource.

Change-Id: Idcd8650f190af901c17933e1210d70f7d34f40fd
Closes-bug: #1689375