Changes to check time since cassandra was last up before starting
If cassandra is down for greater than gc_grace_seconds it should
not be started to avoid deleted data from showing up. This check
was done in /etc/init.d/contrail-database but with move to
systemd and using .service file, we are moving this check to
a new script - contrail-cassandra-check-start which will be
called from ExecStartPre in the contrail-database.service file.
It will return 0 if cassandra can be started, -1 otherwise.

Change-Id: I958141558867cbc5e7bfc1cbc33d35b820f73c12
Partial-Bug: #1664802