Fix VM-Vrouter ref update race.
When there is an add followed by delete, back-to-back, it is possible
VNC layer may not have updated the VM with vrouter info, as that ref
update is a seperate event.

Since VM's are always expected to vrouter info, if the ref is missing,
update that vrouter record in config db by reading from vnc.

Change-Id: I1771bad79ea6f58a768d9481224456fe7089534e
Closes-Bug: #1689905