contrail-status does not show correct status for contrail-database
Seperate out initd_svc and check_return_code variables in
check_svc(). Cassandra is run as initd service even when
the init system used is upstart (ub14.04 bare metal),
supervisor (ub14.04 docker) or sysv and the service status
command always returns the correct status and so we need to
check the return code in all cases for cassandra.

Change-Id: Ibdca3f009ed46c687a949c73884c2222107629ae
Closes-Bug: #1692069