Do not close existing connection in non-GR-Helper case upon new connection
For both BGP and XMPP Sessions, if a new connection request is processed while
a current one is already present and established, drop the new connection
request and retain the old one intact.

This is how base BGP RFC defines when state is established. However, when GR
Helper mode is in effect, switch to current behavior which is to close existing
connection gracefully and process the new connection request. This usually
happens in case of cold reboots of the client machines.

Also re-organize XmppStateMachine::ProcessStreamHeaderMessage() for better
read-ability.

Change-Id: Icdcd9f42bdd5046b1b5bfb07dc0a5bc3472c3a5e
Closes-Bug: #1694789