Ubuntu 1604 DPDK
Have an option not to show services which are not existing in the system.
Currently if there are services which are not existing, contrail-status
shows that it's inactive and disabled on boot.
It was unnecessarily showing DPDK status as inactive on kernel computes
even though kernel compute does not have DPDK. This creates confusion.
So have an option not to show the status in those cases.

Change-Id: I3aeffb122feffdb12ab4521596d414d45c89d4ce
Closes-Bug: 1693097