Implement pagination for collection in VNC API.
Add support for marker and count pagination parameters
in list API upon which collection is returned in batches.
A marker is returned back in every batch which is expected
to be sent back to get the next one. Marker is generated
by db layer (for efficient location of next batch) and
annotated by higher layers if needed (e.g. for sending
'shared' elements of collection).

Closes-Bug: #1433304
Change-Id: I4f728696c898cccdc339a1ec066ced14f7668558