Fix for ip-address not assigned to container on respawn
If a container fails (due to CNI failure or otherwise), kubelet will
delete the failed container and create a new one. In some cases kubelet
is calling CNI multiple times for failed container. We create
interface with names based on POD-UUID. So, both new and old container
will map to same tap interface name. If delete of the container comes
after new container is spawned, the delete must be ignored.

When new container is created, the config file stored by vrouter is
updated with new container-id. Compare the container-id in in this
instance with one present in the config file. Ignore the request if
they do not match. No message is sent to Agent and the config-file is
not deleted

Change-Id: I43f85e49f5304075e8ce32b13295e7cded752172
Closes-Bug: #1694317
(cherry picked from commit 935075945b0ba0fc7a0941beb59c083efc877bbc)