Change test variable from static to non-static
As per review comments for last commit, change a test variable to be dynamic.
Also add couple of new tests for XmppStateMaachine. GR test is disabled due to
the way test runs (no new xmpp connections are formed on new connections unlike
production code). Needs a little more work in the test for the disabled test
to pass

Change-Id: Id7480003819dc241cf6a13af484f603a01b37382
Closes-Bug: #1694789