Load re-balance of collector node after fail-over
Auto re-balance for load distribution is not supported. To achieve fair
load-balance when a node goes DOWN/UP on SIGHUP we will reconnect to
original even if the list is not modified ensuring if nodes as UP, the
subscribers will connect to original and hence achieve a fair load-balance.
(Thnaks to Sundar to suggest the solution)

Change-Id: I91e4db5712726bb027d38674a0370671cd66b552
Closes-Bug:1694793
(cherry picked from commit 4dc34965352373cd20c905112c69ae215dbd342b)