BUM tree subscription skipped.
With the introduction of same RD for Tor agent for vrf, change of vnid was not
handled. In cases where a vn was added with vnid as not set or 0 and then
updated later with non zero value, vrf was not notified.
This used to result in skipping of notify registeration for the problematic vrf.
In turn subscriptions were missed.

Solution:
Handle vnid change.

Change-Id: I7ea7332eb1e64cb0e534f992ef5c8680a54b0313
Closes-bug: #1692795
(cherry picked from commit 8a922e38bfd1e6b2a720d4293fcdf87002cb5a26)