Ubuntu 1604 DPDK
Have an option not to show services which are not existing in the system.
Currently if there are services which are not existing, contrail-status
shows that it's inactive and disabled on boot.
It was unnecessarily showing DPDK status as inactive on kernel computes
even though kernel compute does not have DPDK. This creates confusion.
So have an option not to show the status in those cases.

Change-Id: I876b16c7d22f18802dc8826cbb3d5628ebfbfebc
Closes-Bug: 1693097, 1693951