Do not store OriginVn attribute in Locally originated multicast route
Currently, OriginVn value is determined by looking at the virtual_network_index
stored inside the routing-instance. Because it is possible that when RI is
looked for the index, there may not be an index assigned yet, we some time end
up attaching "0" as the originVn. This is incorrect and can cause issues during
route imports and policy evaluations.

Similar to how other routes handled such as evpn, do not attach OriginVn
attribute at all to the locally originated primary ermvpn route. Instead,
evaluated based on route-target and attach to the secondary routes during
replication.

Change-Id: I55c894f044b0b3bce537b9ace94cd836bfef0fa5
Closes-Bug: #1696853