Agent crash @KSyncObject::CreateImpl
Issue:
Change is requested for ksync entry whose object(to which this entry belong) has
been marked for deletion.
In the bug it was in the code where vxlan id resource is allocated for a logical
switch. Multiple LS can contest for same resource and hence a list is kept of
potential owner of resource plus the current owner. When current owner is being
deleted ownership is passed on to other member in list which in turn issue
change for the new owner of resource. However these deletion can be happening
when ovsdb object is itself marked for deletion. No add/change of any entry is
expected on deleted object.

Solution:
Dont issue add/change on ksync entry in vxlan resource when object is marked for
deletion.

Change-Id: Ib3581a022555d491469217f10a0850c6a5a1d350
Closes-bug: #1699561
(cherry picked from commit 0e369013632d7b9ee92d1383116311334568dac3)